Module Ppx_helpers.Ox

include module type of struct include Modes_lib end

type monadicity =

| Comonadic

| Monadic

val sexp_of_monadicity : monadicity -> Sexplib0.Sexp.t

Modal axes

type locality =

| Global

| Local

val sexp_of_locality : locality -> Sexplib0.Sexp.t

type portability =

| Portable

| Shareable

| Nonportable

val sexp_of_portability : portability -> Sexplib0.Sexp.t

type contention =

| Uncontended

| Shared

| Contended

val sexp_of_contention : contention -> Sexplib0.Sexp.t

type statefulness =

| Stateless

| Observing

| Stateful

val sexp_of_statefulness : statefulness -> Sexplib0.Sexp.t

type visibility =

| Read_write

| Read

| Immutable

val sexp_of_visibility : visibility -> Sexplib0.Sexp.t

type linearity =

| Many

| Once

val sexp_of_linearity : linearity -> Sexplib0.Sexp.t

type uniqueness =

| Unique

| Aliased

val sexp_of_uniqueness : uniqueness -> Sexplib0.Sexp.t

type yielding =

| Unyielding

| Yielding

val sexp_of_yielding : yielding -> Sexplib0.Sexp.t

type forkable =

| Forkable

| Unforkable

val sexp_of_forkable : forkable -> Sexplib0.Sexp.t

type staticity =

| Static

| Dynamic

val sexp_of_staticity : staticity -> Sexplib0.Sexp.t

Nonmodal jkind axes (soon to be nontrivialities)

type externality =

| External_

| External64

| Internal

val sexp_of_externality : externality -> Sexplib0.Sexp.t

type nullability =

| Non_null

| Maybe_null

val sexp_of_nullability : nullability -> Sexplib0.Sexp.t

type separability =

| Non_float

| Separable

| Maybe_separable

val sexp_of_separability : separability -> Sexplib0.Sexp.t

type !'a axis =

| Locality : locality axis

| Portability : portability axis

| Contention : contention axis

| Statefulness : statefulness axis

| Visibility : visibility axis

| Linearity : linearity axis

| Uniqueness : uniqueness axis

| Yielding : yielding axis

| Forkable : forkable axis

| Staticity : staticity axis

val sexp_of_axis : 'a. 'a axis -> Sexplib0.Sexp.t

type !'a nonmodal_axis =

| Externality : externality nonmodal_axis

| Nullability : nullability nonmodal_axis

| Separability : separability nonmodal_axis

type !'a jkind_axis =

| Modal of 'a axis

| Nonmodal of 'a nonmodal_axis

module type Wrapper = Modes_lib.Wrapper

module type Lattice = Modes_lib.Lattice

module type Per_axis = Modes_lib.Per_axis

module type Nonmodal_axis = Modes_lib.Nonmodal_axis

module Base_sort = Modes_lib.Base_sort

module Mode = Modes_lib.Mode

module Modality = Modes_lib.Modality

module Crossing = Modes_lib.Crossing

module Modal = Modes_lib.Modal

module Nonmodal = Modes_lib.Nonmodal

module Jkind_mod = Modes_lib.Jkind_mod

module Layout = Modes_lib.Layout

module type Axis = Modes_lib.Axis

module Axis = Modes_lib.Axis

module Nonmodal_axis = Modes_lib.Nonmodal_axis

module Jkind_axis = Modes_lib.Jkind_axis

Modal axes

module Locality = Modes_lib.Locality

module Portability = Modes_lib.Portability

module Contention = Modes_lib.Contention

module Statefulness = Modes_lib.Statefulness

module Visibility = Modes_lib.Visibility

module Linearity = Modes_lib.Linearity

module Uniqueness = Modes_lib.Uniqueness

module Yielding = Modes_lib.Yielding

Nonmodal jkind axes (soon to be nontrivialities)

module Externality = Modes_lib.Externality

module Nullability = Modes_lib.Nullability

module Separability = Modes_lib.Separability

module Optional = Modes_lib.Optional

module Modalities = Modes_lib.Modalities

module Crossings = Modes_lib.Crossings

module Nonmodals = Modes_lib.Nonmodals

module Jkind_modifiers = Modes_lib.Jkind_modifiers

module Kind = Modes_lib.Kind

val apply_modalities : Modalities.t -> Modes_lib.Modes.t -> Modes_lib.Modes.t

val cross : Crossings.t -> Modes_lib.Modes.t -> Modes_lib.Modes.t

cross crossing mode strengthens mode as it would on a variable of kind _ mod crossing.

In the compiler, this is known as cross_left.

val uncross : Crossings.t -> Modes_lib.Modes.t -> Modes_lib.Modes.t

uncross crossing mode is the weakest mode mode' such that cross crossing mode' is at least as strong as mode. It acts as a sort of inverse to cross.

In the compiler, this is known as cross_right.

module Modes : sig ... end

val of_ast_modes : Ppxlib_jane.modes -> Modes.t

val to_ast_modes : ?simplify:bool -> Modes.t -> loc:Ppxlib.location -> Ppxlib_jane.modes

val of_ast_modalities : mutable_implied:bool -> Ppxlib_jane.modalities -> Modalities.t

val to_ast_modalities : ?simplify:bool -> mutable_implied:bool -> Modalities.t -> loc:Ppxlib.location -> Ppxlib_jane.modalities

val extract_modalities_from_label_declaration : ?handle_globalized:bool -> Ppxlib.label_declaration -> Modalities.t

val extract_modalities_from_construct_arguments : ?handle_globalized:bool -> Ppxlib_jane.Shim.Pcstr_tuple_arg.t -> Modalities.t

val deriving_arg_modes : Modes.t list option Ppxlib.Deriving.Args.param

@@deriving my_ppx ~modes:[ local; shared ] produces

  [ [ Global; Uncontended ]
  ; [ Local; Uncontended ]
  ; [ Global; Shared ]
  ; [ Local; Shared ]
  ]

val deriving_arg_localize : Modes.t list option Ppxlib.Deriving.Args.param

@@deriving my_ppx ~localize produces [[ Local ]]

module Attr : sig ... end